% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrMixture.R
\docType{package}
\name{rrMixture}
\alias{rrMixture}
\title{rrMixture: Reduced-Rank Mixture Models.}
\description{
The rrMixture package provides three important functions currently:
rrmix, rrmix.sim.norm, and initialize.para.
}
